% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_flux.R
\name{calcule_flux}
\alias{calcule_flux}
\title{calcule_flux}
\usage{
calcule_flux(
  analyses,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit,
  col_dates_debit = "date_obs_elab",
  col_debits = "resultat_obs_elab",
  methode = "M6",
  date_debut_calcul,
  date_fin_calcul
)
}
\arguments{
\item{analyses}{dataframe with the concentrations required for calculus. Concentration in mg/L}

\item{col_dates_anal}{Name of the column with date of measurement of concentrations. By default : DatePrel}

\item{col_analyses}{Name of the column with values of concentrations. RsAna by default.}

\item{debit}{dataframe with daily mean flow rate in L/s}

\item{col_dates_debit}{Name of the column with the date of flow rate. Default : date_obs_elab}

\item{col_debits}{Name of the column with the flowrate values. Default : resultat_obs_elab}

\item{methode}{Code of the method for calculus according https://revue-set.fr/article/view/6291 (default : "M6")}

\item{date_debut_calcul}{Date of begining of calculus (for instance "2010-10-01" to start calculus from 1st of october 2010)}

\item{date_fin_calcul}{Date of ending of calculus (for instance "2011-09-30" to end calculus to 30 of september 2011)}
}
\value{
Flux for all the dataset in kg / duration of the measurement (i.e. max(date of flow rate) - min(date of flow rate))
}
\description{
Calculate flux with one of the methods lists in https://revue-set.fr/article/view/6291
}
\details{
M1 : Product of means of sampled Ci and Qi (Preston et al., 1989)
M2 : Mean of instantaneous fluxes : Fi = Ci x Qi (Preston et al., 1989)
M3 : Constant concentration hypothesis around sample (Meybeck et al., 1994)
M4 : Product of means of sampled Ci and annual discharge Q (Shih et al., 1994)
M5 :Flow-weighted mean concentration (Littlewood, 1992)
M6 : Linear interpolation of C (Moatar and Meybeck, 2005)
M9 : Méthode M2 corrigée par un facteur multiplicatif prenant en compte les débits aux instants où l’on n’effectue pas de prélèvements (Cooper, 2004)

Dates date_debut_calcul and date_fin_calcul are included in calculus.
if more than one result of analysis or flow rate is available for a same day, then the mean of all non-NA values is used.
}
\examples{

date_debut_calcul<-"2007-01-01"\%>\%as.Date()
date_fin_calcul<-"2007-12-31"\%>\%as.Date()
  
datafile <- system.file("nitrates.csv", package = "calculeflux")
nitrates <- read.csv2(datafile, encoding = "UTF-8")
nitrates$DatePrel<-nitrates$DatePrel\%>\%as.Date()

datafile <- system.file("debit.csv", package = "calculeflux")
debit <- read.csv2(datafile, encoding = "UTF-8")
debit$date_obs_elab<-debit$date_obs_elab\%>\%as.Date()



calcule_flux(
  analyses=nitrates,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit=debit,
  methode = "M6",
  date_debut_calcul=date_debut_calcul,
  date_fin_calcul=date_fin_calcul
)

methodes<-c("M1", "M2", "M3", "M4", "M5", "M6", "M9")

tmp<-lapply(methodes,
function(x){calcule_flux(
  nitrates,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit,
  methode = x,
  date_debut_calcul=date_debut_calcul,
  date_fin_calcul=date_fin_calcul
)}
)


tmp<-unlist(tmp)
tmp<-data.frame(methode=methodes, flux=tmp/1000)

library(ggplot2)

ggplot2::ggplot(tmp, ggplot2::aes(methode, flux)) + 
  ggplot2::geom_bar(stat="identity") + ggplot2::ylab("Flux en kg NO3/an")

# calcul du flux annuel méthode M6

calcul_flux_annuel<-function(annee)
{
  calcule_flux(
  nitrates,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit,
  methode = "M6",
  date_debut_calcul=paste0(annee,"-10-01")\%>\%as.Date(),
  date_fin_calcul=paste0((annee+1),"-10-01")\%>\%as.Date()
)
 
}

annees<-seq(2002,2021)
tmp<-lapply(annees, calcul_flux_annuel)

tmp<-unlist(tmp)
flux_vilaine_bdhydro<-data.frame(annee_hydro=paste0(annees, " - ", annees+1), flux=tmp/1000)


ggplot2::ggplot(flux_vilaine_bdhydro, ggplot2::aes(annee_hydro, flux)) +
  ggplot2::geom_bar(stat="identity") +
  ggplot2::ylab("flux (t NO3/an)") + 
  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, vjust=0.5)) +
  ggplot2::ggtitle("Flux de nitrates calcul\u00e9s sur la Vilaine \u00e0 Rieux",
          subtitle="NO3 station 04216000 / station hydrom\u00e9trique J930061101")


# calcul du flux bimensuels pour chaque année


library(zoo)
DateSeq_fin_mois <- function(st, en, freq) {
  st <- as.Date(as.yearmon(st)) 
  en <- as.Date(as.yearmon(en)) 
  as.Date(as.yearmon(seq(st, en, by = paste(as.character(12/freq), "months"))), frac = 1)

}

DateSeq_debut_mois <- function(st, en, freq) {
  st <- as.Date(as.yearmon(st))
  en <- as.Date(as.yearmon(en)) 
  as.Date(as.yearmon(seq(st, en, by = paste(as.character(12/freq), "months"))), frac = 0)

}

# # dates de chaque fin de mois
# DateSeq_fin_mois(as.Date("2002-01-01"),as.Date("2022-09-30"),12)

# dates de debut de mois, un mois sur 2
dates_debut_mois<-DateSeq_debut_mois(as.Date("2002-10-01"),as.Date("2022-08-01"),6)

# dates des fins de mois 1 mois sur 2
dates_fin_mois<-DateSeq_fin_mois(as.Date("2002-11-30"),as.Date("2022-09-30"),6)

# mois periode 
mois_periode<-paste0(format(dates_debut_mois, "\%b"),"-",format(dates_fin_mois, "\%b"))
mois_periode<-mois_periode\%>\%factor(levels=unique(mois_periode))

# annnee hydro
an_hydro_periode<-ifelse(format(dates_debut_mois, "\%m")\%>\%as.numeric()>=10,
                         format(dates_debut_mois, "\%Y")\%>\%as.numeric(),
                         format(dates_debut_mois, "\%Y")\%>\%as.numeric()-1)

an_hydro_periode<-paste0(an_hydro_periode, " - ", an_hydro_periode+1)

periodes<-data.frame(debut=dates_debut_mois,
                     fin=dates_fin_mois,
                     mois_periode=mois_periode,
                     an_hydro_periode=an_hydro_periode)


tmp<-lapply(seq(1,nrow(periodes)),
function(x){
  calcule_flux(
  nitrates,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit,
  methode = "M6",
  date_debut_calcul=periodes$debut[x],
  date_fin_calcul=periodes$fin[x]
)
  }
)


flux<-unlist(tmp)
resultats<-cbind(periodes, flux)



ggplot2::ggplot(resultats, ggplot2::aes(mois_periode, flux)) +
  ggplot2::geom_bar(stat="identity") +
  ggplot2::ylab("Flux en kg NO3/an") + 
  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, vjust=0.5)) +
  ggplot2::facet_wrap(an_hydro_periode~.)+
  ggplot2::ggtitle("Flux de nitrates calcul\u00e9s sur la Vilaine \u00e0 Rieux",
          subtitle="NO3 station 04216000 / station hydrom\u00e9trique J930061101")




}
