% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcule_hydraulicite.R
\name{calcule_hydraulicite}
\alias{calcule_hydraulicite}
\title{Calculus of hydraulicity}
\usage{
calcule_hydraulicite(
  debit,
  col_dates_debit = "date_obs_elab",
  col_debits = "resultat_obs_elab",
  date_debut_calcul,
  date_fin_calcul
)
}
\arguments{
\item{debit}{dataframe with daily mean flow rate in m3/s}

\item{col_dates_debit}{Name of the column with the date of flow rate. Default : date_obs_elab}

\item{col_debits}{Name of the column with the flowrate values. Default : resultat_obs_elab}

\item{date_debut_calcul}{Date of begining of calculus (for instance "2010-10-01" to start calculus from 1st of october 2010)}

\item{date_fin_calcul}{Date of ending of calculus (for instance "2011-09-30" to end calculus to 30 of september 2011)}
}
\value{
value of hydraulicity for the period (numeric)
}
\description{
Calculus hydraulicity (Ha) for A period (by default : a = an hydrological year)
}
\details{
Ha = Qa / Qmean
Qa = total flow during period A
Qmean = medium flow on period A over years

Ha < 1 mean tha the period is dry whereas Ha > 1 mean that period is wet.
}
\examples{

date_debut_calcul<-"2022-09-15"\%>\%as.Date()
date_fin_calcul<-"2022-09-30"\%>\%as.Date()

datafile <- system.file("debit.csv", package = "calculeflux")
debit <- read.csv2(datafile, encoding = "UTF-8")
debit$date_obs_elab<-debit$date_obs_elab\%>\%as.Date()


calcule_hydraulicite(debit,
                     col_dates_debit="date_obs_elab",
                     col_debits="resultat_obs_elab",
                     date_debut_calcul,
                     date_fin_calcul)

# example 2

datafile <- system.file("debit.csv", package = "calculeflux")
debit <- read.csv2(datafile, encoding = "UTF-8")
debit$date_obs_elab<-debit$date_obs_elab\%>\%as.Date()

date_debut_calcul=paste0(seq(2002,2021),"-10-01")\%>\%as.Date
date_fin_calcul=paste0(seq(2003,2022),"-09-30")\%>\%as.Date

annees_hydro<-data.frame(date_debut_calcul=date_debut_calcul, 
                         date_fin_calcul=date_fin_calcul,
                         nom=paste0(seq(2002,2021),
                                    "-",
                                    seq(2003,2022)))



tmp<-lapply(seq(2002:2021),
function(x){calcule_hydraulicite(debit,
                     date_debut_calcul=annees_hydro$date_debut_calcul[x],
                     date_fin_calcul=annees_hydro$date_fin_calcul[x])
  }
)


tmp<-unlist(tmp)
tmp<-data.frame(annee=annees_hydro$nom, hydraulicite=tmp)


library(ggplot2)

ggplot2::ggplot(tmp, ggplot2::aes(annee, hydraulicite)) +
  ggplot2::geom_bar(stat="identity") +
  ggplot2::ylab("hydraulicite") + 
  ggplot2::geom_hline(yintercept=1)+
  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, vjust=0.5)) +
  ggplot2::ggtitle("Hydraulicit\u00e9 sur la Vilaine \u00e0 Rieux",
          subtitle="station hydrom\u00e9trique J930061101")



}
