% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculeflux-package.R
\docType{package}
\name{calculeflux-package}
\alias{calculeflux}
\alias{calculeflux-package}
\title{calculeflux: Package To Calculate Nitrogen And Phosphorus Fluxes In Streams}
\description{
Calculation of fluxes with differents methods described in https://revue-set.fr/article/view/6291.
}
\author{
\strong{Maintainer}: Anthony DE BURGHRAVE \email{anthony.deburghrave@eaux-et-vilaine.bzh}

Authors:
\itemize{
  \item Anthony DE BURGHRAVE \email{anthony.deburghrave@eaux-et-vilaine.bzh}
  \item Cédric BRIAND
}

Other contributors:
\itemize{
  \item ETPB Eaux & Vilaine [copyright holder]
}

}
\keyword{internal}
