% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_bv_amont.R
\name{importe_BV_amont}
\alias{importe_BV_amont}
\title{importe_BV_amont}
\usage{
importe_BV_amont(X, Y)
}
\arguments{
\item{X}{: coordonnée du point à partir duquel on souhaite estimer les débits en Lambert 93}

\item{Y}{:  coordonnée du point à partir duquel on souhaite estimer les débits en Lambert 93}
}
\value{
sf object with watershed
}
\description{
Function to calculate watershed using GEOSAS API MNT Surf
https://geosas.fr/web/?page_id=5877
}
\examples{
shp<-importe_BV_amont(X=381425.6,
                      Y=6755598)

plot(sf::st_geometry(shp))

}
