% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_debit.R
\name{importe_debit}
\alias{importe_debit}
\title{importe_debit}
\usage{
importe_debit(X, Y, date_debut, date_fin, source = "dreal")
}
\arguments{
\item{X}{: coordonnée du point à partir duquel on souhaite estimer les débits en Lambert 93}

\item{Y}{:  coordonnée du point à partir duquel on souhaite estimer les débits en Lambert 93}

\item{date_debut}{: date de début de la requête (format date)}

\item{date_fin}{: date de fin de la requête (format date)}

\item{source}{: source des données : archives_dreal ou hub_eau}
}
\value{
dataframe avec 1 colonne date et 1 colonne qjm, qjm étant le débit journalier exprimé en L/s
}
\description{
Fonction pour importer des données de débit à partir des API SIMFEN
https://geosas.fr/web/?page_id=3804
La fonction exploite en priorité la base des archives DREAL et en complément celles hubeau
}
\examples{
test_debit<-importe_debit(X=381425.6, 
                          Y=6755598, 
                          date_debut=as.Date("2010-01-01"), 
                          date_fin = as.Date("2010-01-17"))

#CALCUL DU FLUX DE NITRATE SUR LE SEMNON
datafile <- system.file("donnees_nitrate_semnon.csv", package = "calculeflux")
nitrates_semnon <- read.csv2(datafile, 
                             encoding = "UTF-8",
                             colClasses=c("factor",
                                          "character",
                                          "numeric"),
                             dec=",")
#date mini et maxi
nitrates_semnon$DatePrel<-as.POSIXct(nitrates_semnon$DatePrel, format="\%d/\%m/\%Y \%H:\%M")
summary(nitrates_semnon$DatePrel)


# import des débits sur le Semnon à Eancé
debit_semnon<-importe_debit(X=381425.6, 
                          Y=6755598, 
                          date_debut=as.Date("2010-10-01"), 
                          date_fin = as.Date("2022-09-30"))



# calcul du flux annuel sur la période
calcul_flux_annuel<-function(annee)
{
  calcule_flux(
  analyses=nitrates_semnon,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit=debit_semnon,
  methode = "M6",
  date_debut_calcul=paste0(annee,"-10-01")\%>\%as.Date(),
  date_fin_calcul=paste0((annee+1),"-10-01")\%>\%as.Date()
)
 
}



annees<-seq(2011,2021)
tmp<-lapply(annees, calcul_flux_annuel)
tmp<-unlist(tmp)
rslt_semnon<-data.frame(annee_hydro=paste0(annees, " - ", annees+1), flux=tmp/1000)


ggplot2::ggplot(rslt_semnon, ggplot2::aes(annee_hydro, flux)) +
  ggplot2::geom_bar(stat="identity") +
  ggplot2::ylab("flux (t NO3/an)") + 
  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, vjust=0.5)) +
  ggplot2::ggtitle("Flux de nitrates calcul\u00e9s sur le Semnon \u00e0 Eanc\u00e9",
          subtitle="NO3 station 04211950 / d\u00e9bits 
          estim\u00e9s \u00e0 l\'aide de l\'API SIMFEN / GEOSAS")



##############################

#CALCUL DU FLUX DE NITRATE SUR LA VILAINE AVEC SIMFEN
datafile <- system.file("nitrates.csv", package = "calculeflux")
nitrates <- read.csv2(datafile, encoding = "UTF-8")


#date mini et maxi
nitrates$DatePrel<-nitrates$DatePrel\%>\%as.Date()
summary(nitrates$DatePrel)


# import des débits sur la Vilaine au niveau de la station hydro
debit_vilaine<-importe_debit(X=314920, 
                          Y=6732475, 
                          date_debut=as.Date("2010-10-01"), 
                          date_fin = as.Date("2022-09-30"))



# calcul du flux annuel sur la période
calcul_flux_annuel<-function(annee)
{
  calcule_flux(
  analyses=nitrates,
  col_dates_anal = "DatePrel",
  col_analyses = "RsAna",
  debit=debit_vilaine,
  methode = "M6",
  date_debut_calcul=paste0(annee,"-10-01")\%>\%as.Date(),
  date_fin_calcul=paste0((annee+1),"-10-01")\%>\%as.Date()
)
 
}



annees<-seq(2010,2021)
tmp<-lapply(annees, calcul_flux_annuel)

tmp<-unlist(tmp)
flux_vilaine_simfen<-data.frame(annee_hydro=paste0(annees, " - ", annees+1), flux=tmp/1000)


ggplot2::ggplot(flux_vilaine_simfen, ggplot2::aes(annee_hydro, flux)) +
  ggplot2::geom_bar(stat="identity") +
  ggplot2::ylab("flux (t NO3/an)") + 
  ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 90, vjust=0.5)) +
  ggplot2::ggtitle("Flux de nitrates calcul\u00e9s sur la Vilaine \u00e0 Rieux",
          subtitle="NO3 station 04216000 / d\u00e9bits 
          estim\u00e9s \u00e0 l\'aide de l\'API SIMFEN / GEOSAS")


}
