% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_debit_dates_dispo.R
\name{importe_debit_dates_dispo}
\alias{importe_debit_dates_dispo}
\title{Function wich retrn available dates from SIMFEN API
https://geosas.fr/web/?page_id=3804}
\usage{
importe_debit_dates_dispo()
}
\value{
dataframe with source of flow data and start and end date of data available in the API
}
\description{
Function wich retrn available dates from SIMFEN API
https://geosas.fr/web/?page_id=3804
}
\examples{
importe_debit_dates_dispo()
}
