% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_debit_hubeau.R
\name{importe_debit_hubeau}
\alias{importe_debit_hubeau}
\title{importe_debit_hubeau}
\usage{
importe_debit_hubeau(
  code_entite,
  code_station,
  date_debut_obs_elab = NULL,
  date_fin_obs_elab = NULL
)
}
\arguments{
\item{code_entite}{character or vector with the code of the hydrometric station
or stations for which we want to retrieve the data (mandatory)}

\item{code_station}{character or vector with the code of the hydrometric station
or stations for which we want to retrieve the data (mandatory)}

\item{date_debut_obs_elab}{start date of observations to retrieve (optional)}

\item{date_fin_obs_elab}{end date of observations to retrieve (optional)}
}
\value{
the function returns a tibble with the code of the hydrometric station,
the date, and the daily average flow in L/s
}
\description{
This function allows for importing available flow data from the Hydrometry
HUB'EAU API (https://hubeau.eaufrance.fr/page/api-hydrometrie).
}
\examples{
test<-importe_debit_hubeau(code_entite="J9300611",
                     date_debut_obs_elab=as.Date("2010-01-01"),
                     date_fin_obs_elab=as.Date("2012-12-31"))
}
