% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_debit_siva.R
\name{importe_debit_siva}
\alias{importe_debit_siva}
\title{Imports flow from the SIVA package and calculates mean daily flow}
\usage{
importe_debit_siva(debut, fin, con)
}
\arguments{
\item{debut}{La date de début format POSIXct ou character "\%Y-\%m-\%d \%H:\%M:\%S}

\item{fin}{La fin format POSIXct ou character "\%Y-\%m-\%d \%H:\%M:\%S}

\item{con}{une connexion pool}
}
\value{
A data frame
}
\description{
This function will load data from the Arzal dam
}
\examples{
# Ask anthony or Cédric for the credential file
tryCatch({
  cred <- yaml::read_yaml("credentials.yml")
}, error = function(e) {
  stop("You need to have the credentials to connect to SIVA")
})

hostmysql. <- cred$hostmysql
pwdmysql.  <- cred$pwdmysql
umysql.    <- cred$usermysql

pool <- pool::dbPool(
  drv = RMySQL::MySQL(),
  dbname = "archive_IAV",
  host = hostmysql.,
  username = umysql.,
  password = pwdmysql.,
  port = 3306
)

importe_debit_siva(debut="2020-01-01 00:00:00",
     fin="2020-01-10 00:00:00",
     con=pool)
}
