% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_piezo_hubeau.R
\name{importe_piezo_hubeau}
\alias{importe_piezo_hubeau}
\title{This function allows for importing available piezometric data from the Piezometrie
HUB'EAU API (https://hubeau.eaufrance.fr/page/api-piezometrie).}
\usage{
importe_piezo_hubeau(code_bss, date_debut = NULL, date_fin = NULL)
}
\arguments{
\item{code_bss}{vector with list of old bss code for piezometre (ex. 04193X0022/S2-6)}

\item{date_debut}{start date of observations to retrieve (optional)}

\item{date_fin}{end date of observations to retrieve (optional)}
}
\value{
the function returns a tibble with the code of the piezotric station,
the date, and the daily piezometric
}
\description{
This function allows for importing available piezometric data from the Piezometrie
HUB'EAU API (https://hubeau.eaufrance.fr/page/api-piezometrie).
}
\examples{
tmp<-importe_piezo_hubeau(code_bss=c("04193X0022/S2-6",
                                     "04518X0045/MSM1",
                                     "04218X0035/F1",
                                     "04514X0018/01",
                                     "04513X0112/OS1",
                                     "03555X6010/PZ1"),
                     date_debut = as.Date("1990-01-01"),
                     date_fin=as.Date("2023-01-01"))




  # https://hubeau.eaufrance.fr/api/v1/niveaux_nappes/chroniques?&date_debut_mesure=2010-01-01&date_fin_mesure=2011-01-01&size=20
}
