% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_stations_hydrometriques_hubeau.R
\name{importe_stations_hydrometriques_hubeau}
\alias{importe_stations_hydrometriques_hubeau}
\title{importe_stations_hydrometriques_hubeau
This function allows for importing available hydrometric stations from the
Hydrometry HUB'EAU API (https://hubeau.eaufrance.fr/page/api-hydrometrie).}
\usage{
importe_stations_hydrometriques_hubeau(emprise)
}
\arguments{
\item{emprise}{object of class sf which define the area of the request}
}
\value{
the function returns a sf object with informations about the hydrometric
station
}
\description{
importe_stations_hydrometriques_hubeau
This function allows for importing available hydrometric stations from the
Hydrometry HUB'EAU API (https://hubeau.eaufrance.fr/page/api-hydrometrie).
}
\examples{
utils::data("sage_vilaine", package="calculeflux")
  test<-importe_stations_hydrometriques_hubeau(emprise=sage_vilaine)

  plot(sage_vilaine)
  plot(test, add=TRUE)
}
