% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_stations_piezo_hubeau.R
\name{importe_stations_piezo_hubeau}
\alias{importe_stations_piezo_hubeau}
\title{This function allows for importing available piezometric data from the Piezometrie
HUB'EAU API (https://hubeau.eaufrance.fr/page/api-piezometrie).}
\usage{
importe_stations_piezo_hubeau(
  code_bss = NULL,
  nouveau_bss = NULL,
  shp_emprise = NULL
)
}
\arguments{
\item{code_bss}{vector with list of old bss code for piezometre (ex. 04193X0022/S2-6)}

\item{nouveau_bss}{vector with list of new bss code for piezometre (ex. BSS001DJDP)}

\item{shp_emprise}{sf object

if both code_bss and nouveau_bss are provided, the function return results only for old bss.
if shp_emprise is provided, code_bss and nouveau_bss are not considerated.}
}
\value{
the function returns a tibble with the description of the piezometric station,
}
\description{
This function allows for importing available piezometric data from the Piezometrie
HUB'EAU API (https://hubeau.eaufrance.fr/page/api-piezometrie).
}
\examples{
 utils::data(sage_vilaine, package="calculeflux")
 piezo_vilaine<-importe_stations_piezo_hubeau(shp_emprise=sage_vilaine)
data_old_bss<-importe_stations_piezo_hubeau(code_bss=c("04193X0022/S2-6",
                                                       "03862X0057/PZ"))
data_new_bss<-importe_stations_piezo_hubeau(nouveau_bss=c("BSS001BHVJ",
                                                       "BSS001DJDP",
    "BSS001ESVY"))

                                 

}
