% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_qualite_hubeau.R
\name{importe_qualite_hubeau}
\alias{importe_qualite_hubeau}
\title{importe_qualite_hubeau}
\usage{
importe_qualite_hubeau(
  code_station,
  code_parametre = NULL,
  code_support = NULL,
  code_fraction = NULL,
  date_debut = NULL,
  date_fin = NULL
)
}
\arguments{
\item{code_station}{character or vector with the code of the measurement(s) station(s)}

\item{code_parametre}{character or vector with the code of required parameter (optional)}

\item{date_debut}{start date of observations to retrieve (optional)}

\item{date_fin}{end date of observations to retrieve (optional)}
}
\value{
the function returns a tibble with the code of the measurement station,
the date, the results, ...
}
\description{
This function allows for importing available flow data from the Hydrometry
HUB'EAU API (https://hubeau.eaufrance.fr/page/api-hydrometrie).
}
\details{
All codes comes from https://www.sandre.eaufrance.fr/Rechercher-un-jeu-de-donnees.
}
\examples{
donnees<-importe_qualite_hubeau(code_station="04211950", 
                                code_parametre = "1340",
                                code_support="3")


}
